Update Readme.md
Corrected the assumption that the bootloader was in protected mode. It's
not, the BIOS leaves us in 16 bit mode. We still need to transition to
protected mode.

Unlike 64 bit mode, we can still access the 32 bit registers in real
mode. This is why we see 'ecx' and friends in qemu's monitor, but we
still don't see 'rax'.